Speed up TestCafe CI by running tests in parallel
Split the 18 independent public-library game tests out of the single
publiclibrary.js file into 4 shard files (sharing helpers via the new
publiclibrary-util.js), then run every TestCafe file as its own matrix
job in both the Chrome and Firefox workflows so each spins up its own
server and runs concurrently instead of one long sequential run.

Also trims per-job overhead: cache npm deps via actions/setup-node,
start the server with npm start instead of the nodemon-based npm run
debug, poll for readiness instead of a flat sleep 5, bump checkout/
setup-node to v4, and cancel superseded runs on new pushes.
