+++ .github/workflows/testcafe-firefox.yml
@@ -6,27 +6,40 @@ on:
   pull_request:
     branches: [ main ]
 
+concurrency:
+  group: testcafe-firefox-${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
 
 jobs:
   build:
     name: Run TestCafe Tests
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        testfile:
+          - editor.js
+          - functions.js
+          - publiclibrary-1.js
+          - publiclibrary-2.js
+          - publiclibrary-3.js
+          - publiclibrary-4.js
     steps:
     - name: Checkout repository
-      uses: actions/checkout@v2
+      uses: actions/checkout@v4
     - name: Install Node.js 18.x
-      uses: actions/setup-node@v1
+      uses: actions/setup-node@v4
       with:
         node-version: 18.x
+        cache: npm
     - name: Install dependencies
       run: npm ci
     - name: Start virtualtabletop server
       run: |
-        npm run debug &
-        sleep 5
-        curl -I http://localhost:8272
+        npm start &
+        timeout 30 bash -c 'until curl -sf http://localhost:8272 -o /dev/null; do sleep 0.5; done'
     - name: Run TestCafe tests
       uses: DevExpress/testcafe-action@latest
       with:
         version: "3.7.3"
-        args: "firefox:headless tests/testcafe"
+        args: "firefox:headless tests/testcafe/${{ matrix.testfile }}"
