+++ tests/testcafe/publiclibrary-3.js
@@ -1,0 +1,12 @@
+import { publicLibraryButtons } from './publiclibrary-util.js';
+import { setupTestEnvironment } from './test-util.js';
+
+setupTestEnvironment();
+
+publicLibraryButtons('Functions - CLICK',  0, 'ac3480f0750e9223d6585256d19aa3ab', [ '7u2q' ]);
+publicLibraryButtons('Functions - ROTATE', 0, '5da18c83415df22ca3fc10aba392359e', [ 'c44c', '9kdj', 'w53c', 'w53c' ]);
+publicLibraryButtons('Functions - SELECT', 2, '3c78aba40f3804b70d4d8139d42dc904', [ 'jkmt1']);
+publicLibraryButtons('Functions - SORT',   1, '1fba3113ce09f517f301238fd0626c94', [
+  'ingw', 'k131', 'cnfu', 'i6yz', 'z394', '0v3h', '1h8o', 'v5ra', 'ingw-copy001', 'k131-copy001', 'cnfu-copy001',
+  'i6yz-copy001', 'z394-copy001', '0v3h-copy001'
+]);
