+++ tests/testcafe/publiclibrary-4.js
@@ -1,0 +1,10 @@
+import { publicLibraryButtons } from './publiclibrary-util.js';
+import { setupTestEnvironment } from './test-util.js';
+
+setupTestEnvironment();
+
+publicLibraryButtons('Functions - SCORE', 0, '88c8ef94e7d34f69bf8d0844acce2dfc', [ 'button2', 'button2', 'button7', 'button15', 'seat9', 'scorePlus', 'button17']);
+publicLibraryButtons('Types - Scoreboard', 2, '5b7dbc6f94827871cf11ac913228d7f3', [ 'button1', 'button2', 'button3', 'button4']);
+
+publicLibraryButtons('Types - Dice', 0, 'c2baa316f0f7e04f64af7701cd41b0a9', [ 'dice3', 'dice4', 'dice5',]);
+publicLibraryButtons('Types - Dice', 3, '92066ae0af20d6e26ef447d024fb740a', [ 'dice2', 'dice1', 'button2', 'dice4']);
